# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardMetrics {
  totalInteractions: number;
  supervisorReviews: number;
  safeResponses: number;
  reviewRequired: number;
  previousFailureDetections: number;
  humanCorrections: number;
  memoriesStored: number;
  repeatedMistakesPrevented: number;
}

export interface ActivityItem {
  id: string;
  type: string;
  title: string;
  detail: string;
  timestamp: string;
}

export interface Policy {
  id: string;
  category: string;
  title: string;
  text: string;
  active: boolean;
}

export interface Memory {
  id: string;
  domain: string;
  category: string;
  situation: string;
  customerRequest: string;
  agentResponse: string;
  failure: string;
  humanCorrection: string;
  correctAction: string;
  outcome: string;
  severity: string;
  timestamp: string;
  retrievalCount: number;
}

export interface Alert {
  id: string;
  timestamp: string;
  customerRequest: string;
  workerResponse: string;
  status: string;
  riskLevel: string;
  reason: string;
  relatedMemories: string[];
  finalOutcome: string;
}

export interface SupervisorResult {
  status: string;
  riskLevel: string;
  policyIssues: string[];
  memoryMatches: string[];
  reason: string;
  recommendedAction: string;
  confidence: number;
}

export interface AgentRunInput {
  /** @minLength 1 */
  request: string;
}

export interface AgentRun {
  id: string;
  request: string;
  category: string;
  workerResponse: string;
  finalResponse: string;
  policies: Policy[];
  memories: Memory[];
  supervisor: SupervisorResult;
  mode: string;
  timestamp: string;
  stages: string[];
}

export interface FeedbackInput {
  interactionId: string;
  decision: string;
  whatWasWrong?: string;
  whatShouldHaveHappened?: string;
  correctResponse?: string;
}

export interface Feedback {
  id: string;
  interactionId: string;
  decision: string;
  memoryCreated: boolean;
  timestamp: string;
}

export interface EvaluationInput {
  includeMemory?: boolean;
}

export type EvaluationResultCasesItem = {
  category: string;
  request: string;
  status: string;
  riskLevel: string;
};

export interface EvaluationResult {
  id: string;
  timestamp: string;
  totalCases: number;
  safeCases: number;
  reviewRequiredCases: number;
  policyIssues: number;
  memoryMatches: number;
  humanCorrections: number;
  repeatedFailureDetections: number;
  cases: EvaluationResultCasesItem[];
}

export interface Settings {
  executionMode: string;
  memoryMode: string;
  llmConfigured: boolean;
  hindsightConfigured: boolean;
}

export interface SettingsUpdate {
  executionMode?: string;
}

export type GetMemoriesParams = {
search?: string;
category?: string;
severity?: string;
sort?: string;
};