# lib/api-zod/src/generated/types/agentRun.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Memory } from './memory';
import type { Policy } from './policy';
import type { SupervisorResult } from './supervisorResult';

export interface AgentRun {
  id: string;
  request: string;
  category: string;
  workerResponse: string;
  finalResponse: string;
  policies: Policy[];
  memories: Memory[];
  supervisor: SupervisorResult;
  mode: string;
  timestamp: string;
  stages: string[];
}