# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Alert {
  id: string;
  timestamp: string;
  customerRequest: string;
  workerResponse: string;
  status: string;
  riskLevel: string;
  reason: string;
  relatedMemories: string[];
  finalOutcome: string;
}