# lib/api-zod/src/generated/types/dashboardMetrics.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardMetrics {
  totalInteractions: number;
  supervisorReviews: number;
  safeResponses: number;
  reviewRequired: number;
  previousFailureDetections: number;
  humanCorrections: number;
  memoriesStored: number;
  repeatedMistakesPrevented: number;
}