# lib/api-zod/src/generated/types/evaluationResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EvaluationResultCasesItem } from './evaluationResultCasesItem';

export interface EvaluationResult {
  id: string;
  timestamp: string;
  totalCases: number;
  safeCases: number;
  reviewRequiredCases: number;
  policyIssues: number;
  memoryMatches: number;
  humanCorrections: number;
  repeatedFailureDetections: number;
  cases: EvaluationResultCasesItem[];
}