# lib/api-zod/src/generated/types/feedbackInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FeedbackInput {
  interactionId: string;
  decision: string;
  whatWasWrong?: string;
  whatShouldHaveHappened?: string;
  correctResponse?: string;
}