# lib/api-zod/src/generated/types/memory.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Memory {
  id: string;
  domain: string;
  category: string;
  situation: string;
  customerRequest: string;
  agentResponse: string;
  failure: string;
  humanCorrection: string;
  correctAction: string;
  outcome: string;
  severity: string;
  timestamp: string;
  retrievalCount: number;
}