# lib/api-zod/src/generated/types/policy.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Policy {
  id: string;
  category: string;
  title: string;
  text: string;
  active: boolean;
}