# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  executionMode: string;
  memoryMode: string;
  llmConfigured: boolean;
  hindsightConfigured: boolean;
}