# lib/api-zod/src/generated/types/supervisorResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SupervisorResult {
  status: string;
  riskLevel: string;
  policyIssues: string[];
  memoryMatches: string[];
  reason: string;
  recommendedAction: string;
  confidence: number;
}